ST Port: Store WindowList as short[9][8] so the text-window table uses half the RAM and has a power-of-two size.

+++ common/windows.cpp
@@ -247,7 +247,7 @@ void Set_More_Off(void)
 int Change_Window(int windnum)
 {
     int oldwindow;
-    int* data;
+    short* data;
 
     oldwindow = Window;
     Window = windnum;
